Clean up to nFAPI and error checking UL_TTI_REQ n_pdus
We seprarated out the large if block in the push/pull
of the UL_TTI_REQ fields to make it easier to read
and debug. Furthermore, we add in AsserFatals before
indexing in UL_TTI_REQ PDU list to ensure we are not
indexing past the end of the array.