Revert "kssb < 16 or we cannot do reestablishment: add assert in DU"
This reverts commit 4b9d57380598150debf31eb0c3130090a502f5ce.

This assertion was introduced as we (wrongfully) thought we needed to
calculate the SSB ARFCN at the CU. Since PBCH might carry one bit, which
is not available at the CU, only certain SSB ARFCNs would have been
possible.

With the availability of the SSB ARFCN through the
MeasurementTimingConfiguration, this is not necessary anymore.