Reformulate section on standalone mode
Since tag 2024.w45/commit fd71b54b7d5c483d1108d5862c6cb04f00c9c574,
standalone mode/--sa is the default and does not exist. Correspondingly,
it has been removed from the documentation.

For users reading recent documentation, but running an older version of
OAI, there might be a mismatch in that --sa is necessary to run OAI, but
not given in the documentation. Provide explanatory text to make users
aware of this.