Fix CU to not contain the pointer for MAC/RLC stats
Segmentation fault occured since the commit ID e87ffa2849c0a23fd6aeda1283a6df4f698c250a,
where we try to copy the pointer which was not even been allocated. Before, we were
copying the address of a NULL pointer and therefore no segmentation fault.

This commit fixes the segmentation fault by checking if the memory was allocated, i.e. should not be
in the case of CU (as it doesn't contain MAC/RLC layers).