O-RU: Prepare ethernet packets for application processing.
This commit adds a set of rte_rings for UPlane packets processing
 - uplane_data_ring - this ring contains buffer pointers to be used for UPlane
   packet metadata
 - uplane_symbol_rings - is a set of rings where each ring hold UPlane packet
   metadata for a single symbol

On reception, metadata buffers are dequeued from the data ring and enqueued on
symbol specific rings. When the O-RU application processes the UPlane packet
it frees the rte_mbuf associated with the packet and returns the metadata buffer
back to the data ring.